♻️ update: セクションコンポーネントを共通化し各画面で使用する
